src/pong/requirements.txt
Django
psycopg2-binary
channels
channels_redis
daphne
djangorestframework
